adding orderring endpoint

+++ src/main/java/pl/zajecia/backendproject/shop/model/dto/ProductDto.java
@@ -3,7 +3,7 @@ package pl.zajecia.backendproject.shop.model.dto;
 import lombok.Value;
 import pl.zajecia.backendproject.shop.model.Product;
 
-public record ProductDto(String name, Long price, Long quantity) {
+public record ProductDto(String name, double price, int quantity) {
     public static ProductDto fromEntity(Product product) {
         return new ProductDto(product.getName(), product.getPrice(), product.getQuantity());
     }
